+++ api/java-darts-api/src/main/java/nl/mauritssilvis/darts/java/api/settings/TableType.java
@@ -1,0 +1,22 @@
+/*
+ * Copyright © 2023 Maurits Silvis
+ * SPDX-License-Identifier: GPL-3.0-or-later
+ */
+
+package nl.mauritssilvis.darts.java.api.settings;
+
+import lombok.ToString;
+
+/**
+ * The type of checkout table.
+ *
+ * @author Maurits Silvis
+ * @since 0.5.0
+ */
+@ToString
+public enum TableType {
+    /**
+     * An ascending checkout table.
+     */
+    ASCENDING
+}